ARM: dts: import and convert w400 dtsi from Linux 5.10-rc2
Beelink GT-King, GT-King Pro, and GS-King-X devices are based on the same
W400 reference design, and in Linux they share the common W400 dtsi with
only compatible and audio differences in board files. For u-boot this is
simplified to a single (no audio) dts based on Linux 5.10-rc2 W400 dtsi.

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>